[fontscan] implement shaping.FontMap
